+++ RuntimeEditor/ScriptableObject/EditorLocalizeData.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+using System.Collections.Generic;
+
+namespace Hananoki {
+	[PreferBinarySerialization]
+	public class EditorLocalizeData : ScriptableObject {
+		public string[] m_Strings;
+		public List<string> m_Keys;
+	}
+}