+++ RuntimeEditor/SharedPreference.cs
@@ -19,6 +19,7 @@ namespace Hananoki.SharedModule {
 	[InitializeOnLoad]
 	[Serializable]
 	public class SettingsEditor {
+		public string LCID;
 		public string language;
 		public Color selectionColor = Color.white;
 
@@ -60,8 +61,9 @@ namespace Hananoki.SharedModule {
 			}
 		}
 
+
 		public static void LoadLocalize() {
-			EditorLocalize.Load( Package.name, i.language, "2ca67e52d0e4c5a439729c95e8bf5e45" );
+			EditorLocalize.Load( Package.name, "95cedfc7731853946b0b3650f175d73a", E.i.LCID );
 
 			if( s_localizeEvent == null ) {
 				s_localizeEvent = new List<MethodInfo>();
@@ -87,8 +89,10 @@ namespace Hananoki.SharedModule {
 					}
 				}
 			}
-			foreach( var m in s_localizeEvent ) m.Invoke( null, null );
-
+			foreach( var m in s_localizeEvent ) {
+				//Debug.Log("Invoke: "+m.Name);
+				m.Invoke( null, null );
+			}
 			HEditorWindow.RepaintWindow( typeof( EditorWindow ) );
 		}
 	}
@@ -105,6 +109,9 @@ namespace Hananoki.SharedModule {
 		}
 		public static List<Lang> localeFiles;
 
+		public static List<EditorLocalize.LCIDString> m_localeNames;
+		public static int m_localeIndex;
+
 		public static void Open() {
 			var w = GetWindow<SettingsEditorWindow>();
 			w.SetTitle( new GUIContent( Package.name, EditorIcon.settings ) );
@@ -118,28 +125,32 @@ namespace Hananoki.SharedModule {
 		public static void DrawGUI() {
 			E.Load();
 
-			if( localeFiles == null ) {
-				var files = DirectoryUtils.GetFileGUIDs( AssetDatabase.GUIDToAssetPath( "95cedfc7731853946b0b3650f175d73a" ), "*.csv" );
-				localeFiles = files
-					//.Select( x => AssetDatabase.GUIDToAssetPath( x ).FileNameWithoutExtension() )
-					.Select( x => new Lang { file = x, popName = (string) EditorLocalize.s_dic2[ AssetDatabase.GUIDToAssetPath( x ).FileNameWithoutExtension() ] } )
-					.ToList();
-				localeFiles = localeFiles.OrderBy( x => x.popName ).ToList();
-				//localeFilesPopup = localeFiles.Select( x => AssetDatabase.GUIDToAssetPath( x ).FileNameWithoutExtension() ).Select( x => (string) EditorLocalize.s_dic2[ x ] ).ToArray();
+
+			if( m_localeNames == null ) {
+				m_localeNames = new List<EditorLocalize.LCIDString>();// EditorLocalize.s_lcidTable.OrderBy( x => x.NAME ).ToList();
+				var files = "95cedfc7731853946b0b3650f175d73a".GetFilesFromAssetGUID().Select( x => x.FileNameWithoutExtension() ).ToList();
+				foreach( var rp in files ) {
+					var p = EditorLocalize.s_lcidTable.Find( x => x.LCID == rp );
+					if( p == null ) continue;
+					m_localeNames.Add( p );
+				}
 			}
-			//if( UnitySymbol.Has( "UNITY_2018_3_OR_NEWER" ) ) {
-			//	EditorGUILayout.LabelField( "Support", "2018.3 - xxx" );
-			//}
-			//else {
-			//	EditorGUILayout.LabelField( "Support", "Not Support" );
-			//}
 
 			EditorGUI.BeginChangeCheck();
-			int n = localeFiles.FindIndex( x => x.file == E.i.language );
-			if( n < 0 ) {
-				n = 0;
+			var idx = m_localeNames.FindIndex( x => x.LCID == E.i.LCID );
+			if( idx < 0 ) {
+				idx = m_localeNames.FindIndex( x => x.LCID == "en-US" );
+
 			}
-			n = EditorGUILayout.Popup( S._Language, n, localeFiles.Select( x => x.popName ).ToArray() );
+			idx = EditorGUILayout.Popup( S._Language, idx, m_localeNames.Select( x => x.NAME ).ToArray() );
+			if( EditorGUI.EndChangeCheck() ) {
+				E.i.LCID = m_localeNames[ idx ].LCID;
+				E.LoadLocalize();
+				E.Save();
+			}
+
+			EditorGUI.BeginChangeCheck();
+
 
 			GUILayout.Space( 4 );
 			try {
@@ -168,17 +179,9 @@ namespace Hananoki.SharedModule {
 
 
 			if( EditorGUI.EndChangeCheck() ) {
-				E.i.language = localeFiles[ n ].file;
-				E.LoadLocalize();
 				E.Save();
 			}
 
-			//EditorGUI.BeginChangeCheck();
-			//var _color = EditorGUILayout.ColorField( nameof( Pref.selectionColor ).nicify(), Pref.i.selectionColor );
-			//if( EditorGUI.EndChangeCheck() ) {
-			//	Pref.i.selectionColor = _color;
-			//	Pref.Save();
-			//}
 			GUILayout.Space( 8f );
 
 			//#if ENABLE_HANANOKI_SETTINGS && LOCAL_TEST